feat(sessionclaims): review feedback
