+++ lib/ts/recipe/emailverification/api/generateEmailVerifyToken.ts
@@ -37,7 +37,7 @@ export default async function generateEmailVerifyToken(
 
     const result = await apiImplementation.generateEmailVerifyTokenPOST({
         options,
-        session,
+        session: session!,
         userContext,
     });
