+++ lib/ts/recipe/passwordless/api/consumeCode.ts
@@ -15,7 +15,6 @@
 
 import { send200Response } from "../../../utils";
 import STError from "../error";
-import EmailVerification from "../../emailverification/recipe";
 import { APIInterface, APIOptions } from "..";
 import { makeDefaultUserContextFromAPI } from "../../../utils";
 
@@ -76,24 +75,6 @@ export default async function consumeCode(apiImplementation: APIInterface, optio
     );
 
     if (result.status === "OK") {
-        if (result.user.email !== undefined) {
-            const emailVerificationInstance = EmailVerification.getInstance();
-            if (emailVerificationInstance) {
-                const tokenResponse = await emailVerificationInstance.recipeInterfaceImpl.createEmailVerificationToken({
-                    userId: result.user.id,
-                    email: result.user.email,
-                    userContext,
-                });
-
-                if (tokenResponse.status === "OK") {
-                    await emailVerificationInstance.recipeInterfaceImpl.verifyEmailUsingToken({
-                        token: tokenResponse.token,
-                        userContext,
-                    });
-                }
-            }
-        }
-
         delete (result as any).session;
     }
